proposed fix for #391
issue #391 is occuring because theme afterInputReady is being invoked
after theme addInputError for a validation error.

the fix involves adding a field to the input object for tracking a
queued error message. addInputError and removeInputError keep this
queued error message up to date as neeeded. afterInputReady invokes
addInputError on a queued error message once it is invoked.

my goals for this fix were (1) to maintain consistency with the way
each theme has a separate implementation for controlgroup ~ controls ~
group, (2) to avoid modifications of control flow around the invokations
of afterInputReady, addInputError, removeInputError.