Added hidden input support.
Example use case: Suppose you store an array of objects. Every property
of these objects should be editable by the client but the requirement
is to specifically identify if an object is modified totally or newly
added to the array. With hidden inputs, you can easily store the id
of the objects and can tell if an object is newly created or if all
the fields of the object is modified.

Signed-off-by: Mehmet Baker <mehmet.baker@gmail.com>